New config options + more.
InitApiUsingExternalPlugin indicates if the API gets initiated by an external plugin, or if it gets initiated by the InteractiveButtons plugin. Also changed the Author and changed some of the config description.

+++ InteractiveButtons/Plugin.cs
@@ -14,13 +14,13 @@ namespace InteractiveButtons
 {
     public sealed class Plugin : Plugin<Config>
     {
-        public override string Author => "KadotCom";
+        public override string Author => "InteractDev (KadotCom)";
 
         public override string Name => "InteractiveButtons";
 
         public override string Prefix => Name;
 
-        public override Version Version => new Version(1,1,0);
+        public override Version Version => new Version(1,2,0);
         public override Version RequiredExiledVersion => new Version(8,2,0);
         public static Plugin Instance;
 
@@ -32,8 +32,14 @@ namespace InteractiveButtons
 
             
             RegisterEvents();
-
-            API.API.Init();
+            if (!Config.InitApiUsingExternalPlugin)
+            {
+                API.API.Init();
+            }
+            else
+            {
+                PluginAPI.Core.Log.Debug("InitApiUsingExternalPlugin is set to true, meaning that the API will need to be initiated by an external plugin.");
+            }
 
             base.OnEnabled();
         }
@@ -42,7 +48,14 @@ namespace InteractiveButtons
         {
             UnregisterEvents();
 
-            API.API.Destroy();
+            if (!Config.InitApiUsingExternalPlugin)
+            {
+                API.API.Destroy();
+            }
+            else
+            {
+                PluginAPI.Core.Log.Debug("InitApiUsingExternalPlugin is set to true, meaning that the API will need to be uninitiated by an external plugin.");
+            }
             
             Instance = null;
